refactor to subtype
